chore(structure): change structure
